Move relevant methods from `partial_snapshot` module to `snapshot` module
