+++ lib/segment/src/entry/snapshot_entry.rs
@@ -21,4 +21,6 @@ pub trait SnapshotEntry {
         manifest: Option<&SegmentManifests>,
         snapshotted_segments: &mut HashSet<String>,
     ) -> OperationResult<()>;
+
+    fn collect_segment_manifests(&self, manifests: &mut SegmentManifests) -> OperationResult<()>;
 }
